nft_bot/requirements.txt
aiohttp==3.8.1
mysql-connector-python==8.0.27
selenium==4.1.0
selenium-wire==4.5.6